\name{ManyTests-package}
\alias{ManyTests-package}
\alias{ManyTests}
\docType{package}
\title{
Multiple Testing Procedures of Cox (2011) and Wong and Cox (2007)
}

\description{
Performs the multiple testing procedures of Cox (2011) and Wong and Cox (2007).
}
\details{
\tabular{ll}{
Package: \tab ManyTests\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-10-30\cr
License: \tab GPL-2\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Christiana Kartsonaki

Maintainer: Christiana Kartsonaki <christiana.kartsonaki@gmail.com>
}

\references{
Cox, D. R. (2011). Discovery: a statistical perspective. \emph{Phystat Conference CERN}. <doi:10.5170/CERN-2011-006>

Cox, D. R. and Wong, M. Y. (2004). A simple procedure for the selection of significant effects. \emph{Journal of the Royal Statistical Society} B \bold{66} (2), 395--400. <doi:10.1111/j.1369-7412.2004.05695.x>

Wong, M. Y. and Cox, D. R. (2007). On the screening of large numbers of significance tests. \emph{Journal of Applied Statistics} \bold{34} (7), 779--783. <doi:10.1080/02664760701240014>
}

%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
